pyproject.toml
[tool.poetry]
name = "bitpin"
version = "0.1.0"
description = "Software Engineer Task: Back-End"
authors = ["Sepehr <bazyar.sepehr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.2"
pre-commit = "^3.7.1"
ruff = "^0.4.7"
pytest = "^8.2.1"
pytest-django = "^4.8.0"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    "migrations",
]

# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP",  # pyupgrade
]
extend-select = ["E501"]
ignore = []


[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "F401",
]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
